v1.2.77 TT365 League Rankings are currentlyy broken on main site
Updated `TeamSummary.razor` to conditionally display the "Ranking" column based on the presence of non-zero league rankings. Adjusted table row rendering to align with this conditional logic.

Refactored `GetTeamStats.cs` to handle changes in column indices due to the removal of the `LeagueRanking` field. Defaulted `LeagueRanking` to `0` as it is no longer tracked. Simplified the logic for extracting "Player of the Match" awards and player form data.